backend/requirements.txt
flask
werkzeug
python-docx
PyPDF2
sklearn
Flask>=2.2.2